mysite/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
kavenegar==1.1.2
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
